update UI of courseCard and courseDetail of student UI

+++ source_code/back-end/online_learning/src/main/java/com/vitube/online_learning/dto/response/CourseStatisticResponse.java
@@ -2,6 +2,8 @@ package com.vitube.online_learning.dto.response;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Builder
@@ -10,5 +12,5 @@ import lombok.*;
 public class CourseStatisticResponse {
     private String title;
     private int totalRegistrations;
-    private float totalEarnings;
+    private BigDecimal totalEarnings;
 }
